move BEFORE_UPSERT into wnprc_triggers.js
